fixed the bug

+++ src/TrMap/TrHeightMap.hpp
@@ -8,7 +8,7 @@ class TrMap;
 // is this too much memory? make sure to check
 class TrHeightMap : public TrMapData<double> {
  public:
-  TrMapData<uint32_t> m_terrace;
+  TrMapData<double> m_terrace;
   TrHeightMap(int rows, int cols)
       : TrMapData<double>(rows, cols), m_terrace(rows, cols) {}
